# internal/server/openapi.yaml
openapi: "3.0.0"
info:
  version: 3.0.0
  title: Peregrine API documentation
servers:
  - url: http://edge.api.peregrine.ga:8080/
  - url: http://api.peregrine.ga:8080/
  - url: http://localhost:8080/
paths:
...
  /refresh:
    post:
      summary: Retrieve a new access token via a refresh token
      operationId: refresh
      tags:
        - authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - refreshToken
              properties:
                refreshToken:
                  $ref: "#/components/schemas/refreshToken"
      responses:
        "200":
          description: Sucessfully generated a new access token
          content:
            application/json:
              schema:
                required:
                  - accessToken
                properties:
                  accessToken:
                    $ref: "#/components/schemas/accessToken"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "422":
          description: Request body syntax was invalid
          content:
            text/plain:
              schema:
                type: string
                example: Unprocessable Entity
        "500":
          $ref: "#/components/responses/internalServerError"
  /users:
    post:
      summary: Create a new user
      description:
        Note that if you specify roles higher than your own roles they will be reset.
        For example, you can't create a verified user in another realm if you're not
        a global admin, and you can't create an admin if you're not an admin.
      operationId: createUser
      security:
        - BearerAuth: []
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        "201":
          description: Successfully created user
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "409":
          $ref: "#/components/responses/conflictError"
        "422":
          $ref: "#/components/responses/unprocessableEntityError"
        "500":
          $ref: "#/components/responses/internalServerError"
    get:
      summary: Get all visible users
      description:
        Note that only admins can get users. Global admins can get all users, and realm
        admins can get all users in their realm.
      operationId: getUsers
      security:
        - BearerAuth: []
      tags:
        - users
      responses:
        "200":
          description: Successfully fetched users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "500":
          $ref: "#/components/responses/internalServerError"
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          $ref: "#/components/schemas/id"
        required: true
        description: Numeric User ID
    get:
      summary: Get a user
      operationId: getUser
      security:
        - BearerAuth: []
      tags:
        - users
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "400":
          $ref: "#/components/responses/badRequestError"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "404":
          $ref: "#/components/responses/notFoundError"
        "500":
          $ref: "#/components/responses/internalServerError"
    patch:
      summary: Update all or parts of a user
      description:
        Any null or undefined fields will not be updated. Global admins can patch any user,
        realm admins can patch users in their realm, and any user can patch themselves.
      operationId: patchUser
      security:
        - BearerAuth: []
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        "204":
          description: Successfully patched user
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "404":
          $ref: "#/components/responses/notFoundError"
        "500":
          $ref: "#/components/responses/internalServerError"
    delete:
      summary: Delete a user
      operationId: deleteUser
      security:
        - BearerAuth: []
      tags:
        - users
      responses:
        "204":
          description: Successfully deleted user
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "404":
          $ref: "#/components/responses/notFoundError"
        "500":
          $ref: "#/components/responses/internalServerError"
...
components:
...
  responses:
    internalServerError:
      description: Failed due to an internal server error
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/internalServerError"
    unauthorizedError:
      description: Invalid access token, please obtain a new access token
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
    forbiddenError:
      description: Your access token is valid, but you lack the roles to perform this operation
      content:
        text/plain:
          schema:
            type: string
            example: Forbidden
    notFoundError:
      description: Unable to find that resource
      content:
        text/plain:
          schema:
            type: string
            example: Not Found
    unprocessableEntityError:
      description: Request body syntax was invalid
      content:
        text/plain:
          schema:
            type: string
            example: Unprocessable Entity
    conflictError:
      description: A resource with a similar unique key exists (e.g. username or id)
      content:
        text/plain:
          type: string
          example: Conflict
    badRequestError:
      description: Your request was invalid
      content:
        text/plain:
          type: string
          example: Bad Request
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    id:
      description: Auto-increment 64-bit integer that identifies a resource
      type: integer
      format: int64
      example: 1
      readOnly: true
...
    user:
      required:
        - id
        - username
        - password
        - realmId
        - firstName
        - lastName
      properties:
        id:
          $ref: "#/components/schemas/id"
        username:
          type: string
          example: franklin
        password:
          type: string
          example: Sxam0dO3aMQW
          writeOnly: true
        realmId:
          $ref: "#/components/schemas/id"
        firstName:
          type: string
          example: Franklin
        lastName:
          type: string
          example: Harding
        stars:
          $ref: "#/components/schemas/stars"
        roles:
          $ref: "#/components/schemas/roles"
    stars:
      type: array
      items:
        type: string
        example: 2018pncmp
    roles:
      required:
        - isSuperAdmin
        - isAdmin
        - isVerified
      properties:
        isSuperAdmin:
          type: boolean
          example: true
        isAdmin:
          type: boolean
          example: true
        isVerified:
          type: boolean
          example: false
    accessToken:
      type: string
      description: Short lived (1 day) JWT for API authorization
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwZXJlZ3JpbmVSb2xlcyI6eyJpc1N1cGVyQWRtaW4iOnRydWUsImlzQWRtaW4iOnRydWUsImlzVmVyaWZpZWQiOnRydWV9LCJwZXJlZ3JpbmVSZWFsbSI6MSwiZXhwIjoxNTQ3NjcyNDE2LCJzdWIiOiIxIn0.FMsGteZSZ53YmBFAKFHTde_vTymvuT0iFfLGPmcbkDQ
    refreshToken:
      type: string
      description: Long lived (4 week) JWT for generating new access tokens without re-authenticating
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwZXJlZ3JpbmVQYXNzd29yZENoYW5nZWQiOiIyMDE5LTAxLTA5VDIyOjUxOjMxWiIsImV4cCI6MTU1MDAwNTIxNiwic3ViIjoiMSJ9.cCnaInxuIdrkZLTkVpx4i9iALoUzQTIYKWc5OtTroiM
    internalServerError:
      type: string
      example: Internal Server Error
...